1.0.0 - Split up cookieManager and unit tests
